Proyectos IA /Pa09 /2-3-coder/requirements.txt
streamlit>=1.18
ultralytics>=8.0
pillow
pandas
plotly
opencv-python-headless
